initial refactor and added dependencies
