+++ .travis.yml
@@ -1,3 +1,16 @@
 language: node_js
+
 node_js:
-  - "6"
+  - 6
+
+env:
+  - CXX=g++-4.8
+addons:
+  apt:
+    sources:
+      - ubuntu-toolchain-r-test
+      - mongodb-3.2-precise
+
+    packages:
+      - g++-4.8
+      - mongodb-org-server
