+++ lib/app.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const app = module.exports = express();
-const router = require('./router');
+const notes = require('./routes/notes');
 
-app.use('/api', router);
+app.use('/api/notes', notes);
 
 //error handling
 app.use((error, req, res, next)=>{
