+++ package.json
@@ -1,13 +1,13 @@
 {
-  "name": "expressjs-single-resource-api",
+  "name": "express-mongo-rest-api",
   "version": "2.1.0",
-  "description": "This is an http server that will act as a data store, backed by in-memory storage objects.",
+  "description": "This is an http server that will act as a data store using MongoDB and Mongoose.",
   "main": "index.js",
   "directories": {
     "test": "test"
   },
   "scripts": {
-    "start": "node index.js",
+    "start": "nodemon index.js",
     "test": "mocha",
     "lint": "eslint **/*.js"
   },
@@ -25,9 +25,11 @@
     "chai": "^3.5.0",
     "chai-http": "^3.0.0",
     "eslint": "^3.1.1",
-    "mocha": "^2.5.3"
+    "mocha": "^2.5.3",
+    "nodemon": "^1.10.0"
   },
   "dependencies": {
-    "express": "^4.14.0"
+    "express": "^4.14.0",
+    "mongoose": "^4.5.7"
   }
 }
